Refactor mcpClient with modular approach
- Added new `capabilities.ts`, `config.ts`, and `connectivity.ts` files to manage MCP server capabilities and configurations.
- Introduced improved OAuth handling with dynamic client registration and enhanced token exchange logic.
- Updated error handling in OAuth processes to provide clearer feedback on authorization issues.
- Implemented caching for server capabilities and resources to optimize performance and reduce redundant requests.
- Enhanced the `exchangeOAuthCode` function to support multiple client identification strategies and fallback mechanisms.
- Added new utility functions for managing prompts and resources in the MCP client.
- Improved overall structure and modularity of the OAuth and MCP client codebase for better maintainability.
