+++ lib/ai/mastra/pkceStore.ts
@@ -2,12 +2,14 @@
  * Redis-backed PKCE store using the shared KV client
  * Ensures 10-minute TTL and one-time retrieval semantics
  */
-import { OAUTH_CONFIG } from './mcpClient';
+import { OAUTH_CONFIG } from './config';
 import { kv } from '@/lib/kv';
 
 interface PKCEData {
 	codeVerifier: string;
 	clientId: string;
+	resource?: string;
+	redirectUri?: string;
 	createdAt: number;
 }
 
@@ -18,10 +20,14 @@ export async function storePKCEVerifier(
 	state: string,
 	codeVerifier: string,
 	clientId?: string,
+	resource?: string,
+	redirectUri?: string,
 ): Promise<void> {
 	const value: PKCEData = {
 		codeVerifier,
 		clientId: clientId || OAUTH_CONFIG.clientName,
+		resource,
+		redirectUri,
 		createdAt: Date.now(),
 	};
 
@@ -32,9 +38,12 @@ export async function storePKCEVerifier(
 	});
 }
 
-export async function retrievePKCEData(
-	state: string,
-): Promise<{ codeVerifier: string; clientId: string } | null> {
+export async function retrievePKCEData(state: string): Promise<{
+	codeVerifier: string;
+	clientId: string;
+	resource?: string;
+	redirectUri?: string;
+} | null> {
 	const key = `${PREFIX}${state}`;
 	const data = await kv.get<PKCEData>(key);
 	if (!data) {
@@ -49,5 +58,10 @@ export async function retrievePKCEData(
 
 	// Best-effort one-time read: delete after retrieval
 	await kv.delete(key);
-	return { codeVerifier: data.codeVerifier, clientId: data.clientId };
+	return {
+		codeVerifier: data.codeVerifier,
+		clientId: data.clientId,
+		resource: data.resource,
+		redirectUri: data.redirectUri,
+	};
 }
